Add dequeue helper for mutex lock
Currently, mo_mutex_lock() will call mutex_block_atomic() to mark the
running task as TASK_BLOCKED so that it won't be selected by the old scheduler.
To support the ready queue consistency that always keeps runnable tasks,
the dequeuing path should be included when mutex_block_atomic() is
called.

This commit adds _sched_blocked_dequeue() helper and will be applied in
mutex_block_atomic() in the following commit.